feat: Implement CronJobManager

+++ internal/repository/cron_job_repository.go
@@ -16,17 +16,17 @@ type CronJobRepository interface {
 	FetchDueCronJobs(ctx context.Context, page int, pageSize int) (*models.PaginationResult[models.CronJob], error)
 
 	// UpdateJobRunTimes updates the LastRunAt and NextRunAt timestamps after execution.
-	UpdateJobRunTimes(ctx context.Context, jobID int, lastRunAt, nextRunAt time.Time) error
+	UpdateJobRunTimes(ctx context.Context, jobID int64, lastRunAt, nextRunAt time.Time) error
 
 	// MarkSuccess marks the job as successfully executed.
-	MarkSuccess(ctx context.Context, jobID int) error
+	MarkSuccess(ctx context.Context, jobID int64) error
 
 	// MarkFailure marks the job execution as failed with the given error message.
-	MarkFailure(ctx context.Context, jobID int, errMsg string) error
+	MarkFailure(ctx context.Context, jobID int64, errMsg string) error
 
 	// Activate Job
-	Activate(ctx context.Context, jobID int) error
+	Activate(ctx context.Context, jobID int64) error
 
 	// Deactivate Job
-	DeActivate(ctx context.Context, jobID int) error
+	DeActivate(ctx context.Context, jobID int64) error
 }
